pyproject.toml
[project]
name = "vibd"
version = "0.1.0"
description = ""
authors = [
    {name = "Until",email = "kostyuchenko.work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
